Consolidate various PID stuff
