+++ ehri-io/src/main/resources/reference.conf
@@ -28,9 +28,14 @@ io {
     ead {
       includeRevisions = false
     }
-    pids {
-      prefix = "ark:/41045/"
-      label = "EHRI Archival Resource Key"
-    }
+  }
+
+  pids {
+    # 41045 is the NAAN, p2 the shoulder
+    prefix = "ark:/41045/p2"
+    # the length of the generated portion
+    length = 8
+    # PID label
+    label = "EHRI Archival Resource Key"
   }
 }
