+++ ehri-io/src/test/java/eu/ehri/project/importers/ead/EadSyncTest.java
@@ -3,6 +3,7 @@ package eu.ehri.project.importers.ead;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import eu.ehri.project.IdGeneratorProvider;
 import eu.ehri.project.definitions.EventTypes;
 import eu.ehri.project.definitions.Ontology;
 import eu.ehri.project.importers.PreImportCallback;
@@ -13,7 +14,7 @@ import eu.ehri.project.models.*;
 import eu.ehri.project.models.base.Accessor;
 import eu.ehri.project.models.base.PermissionScope;
 import eu.ehri.project.models.events.SystemEvent;
-import eu.ehri.project.models.idgen.ArkIdGenerator;
+import eu.ehri.project.models.idgen.RandomIdGenerator;
 import eu.ehri.project.persistence.Bundle;
 import org.junit.Before;
 import org.junit.Test;
@@ -31,7 +32,7 @@ public class EadSyncTest extends AbstractImporterTest {
 
     private Repository repo;
     private SaxImportManager importManager;
-    private final ArkIdGenerator arkIdGenerator = ArkIdGenerator.create(10);
+    private static final RandomIdGenerator arkIdGenerator = IdGeneratorProvider.getIdGenerator();
 
     @Before
     public void setUp() throws Exception {
