+++ ehri-ws/src/main/java/eu/ehri/project/ws/AdminResource.java
@@ -27,9 +27,6 @@ import com.tinkerpop.blueprints.Graph;
 import com.tinkerpop.blueprints.Vertex;
 import com.tinkerpop.blueprints.util.io.graphson.GraphSONMode;
 import com.tinkerpop.blueprints.util.io.graphson.GraphSONWriter;
-import eu.ehri.project.models.idgen.ArkIdGenerator;
-import eu.ehri.project.models.idgen.RandomIdGenerator;
-import eu.ehri.project.ws.base.AbstractResource;
 import eu.ehri.project.acl.PermissionType;
 import eu.ehri.project.acl.wrapper.AclGraph;
 import eu.ehri.project.core.Tx;
@@ -40,14 +37,10 @@ import eu.ehri.project.models.UserProfile;
 import eu.ehri.project.models.base.Accessor;
 import eu.ehri.project.persistence.Bundle;
 import eu.ehri.project.tools.JsonDataExporter;
+import eu.ehri.project.ws.base.AbstractResource;
 import org.neo4j.graphdb.GraphDatabaseService;
 
-import javax.ws.rs.Consumes;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
+import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
